test(pin-input): recategorize and clean up browser tests
