+++ packages/react/src/components/pin-input/pin-input.test.browser.tsx
@@ -1,91 +1,13 @@
-import { a11y, page, render, renderHook } from "#test/browser"
+import { page, render } from "#test/browser"
 import { PinInput } from "."
-import { usePinInput } from "./use-pin-input"
 
 const getInputs = () =>
-  page.getByRole("textbox").elements() as HTMLInputElement[]
+  page
+    .getByRole("textbox")
+    .elements()
+    .filter((el) => el instanceof HTMLInputElement)
 
 describe("<PinInput />", () => {
-  test("renders component correctly", async () => {
-    await a11y(<PinInput.Root />)
-  })
-
-  test("sets `displayName` correctly", () => {
-    expect(PinInput.Root.name).toBe("PinInputRoot")
-    expect(PinInput.Field.name).toBe("PinInputField")
-  })
-
-  test("sets `className` correctly", async () => {
-    await render(
-      <PinInput.Root data-testid="root">
-        <PinInput.Field data-testid="field" index={0} />
-      </PinInput.Root>,
-    )
-
-    await expect
-      .element(page.getByTestId("root"))
-      .toHaveClass("ui-pin-input__root")
-    await expect
-      .element(page.getByTestId("field"))
-      .toHaveClass("ui-pin-input__field")
-  })
-
-  test("merges root className with caller className", async () => {
-    await render(<PinInput.Root className="from-caller" data-testid="root" />)
-
-    await expect
-      .element(page.getByTestId("root"))
-      .toHaveClass("ui-pin-input__root")
-    await expect.element(page.getByTestId("root")).toHaveClass("from-caller")
-  })
-
-  test("merges getRootProps values with caller props", async () => {
-    const restOnClick = vi.fn()
-    const callerOnClick = vi.fn()
-    const { result } = await renderHook(() =>
-      usePinInput({
-        className: "from-rest",
-        style: { backgroundColor: "red", paddingTop: "4px" },
-        onClick: restOnClick,
-      }),
-    )
-
-    const rootProps = result.current.getRootProps({
-      className: "from-caller",
-      style: { color: "blue", paddingTop: "8px" },
-      onClick: callerOnClick,
-    })
-
-    expect(rootProps.className).toContain("from-rest")
-    expect(rootProps.className).toContain("from-caller")
-    expect(rootProps.style).toMatchObject({
-      backgroundColor: "red",
-      color: "blue",
-      paddingTop: "8px",
-    })
-    expect(rootProps.role).toBe("group")
-
-    rootProps.onClick?.({} as any)
-
-    expect(restOnClick).toHaveBeenCalledTimes(1)
-    expect(callerOnClick).toHaveBeenCalledTimes(1)
-  })
-
-  test("renders the correct number of input elements", async () => {
-    await render(<PinInput.Root items={5} />)
-    expect(getInputs()).toHaveLength(5)
-  })
-
-  test("id prop applies correctly", async () => {
-    const testId = "test"
-    const { container } = await render(<PinInput.Root id={testId} />)
-
-    expect(container.querySelector(`#${testId}`)).toBeInTheDocument()
-    expect(container.querySelector(`#${testId}-1`)).toBeInTheDocument()
-    expect(container.querySelector(`#${testId}-2`)).toBeInTheDocument()
-    expect(container.querySelector(`#${testId}-3`)).toBeInTheDocument()
-  })
-
   test('allows alphanumeric input when type is "alphanumeric"', async () => {
     const { user } = await render(
       <>
@@ -129,61 +51,6 @@ describe("<PinInput />", () => {
     expect(handleComplete).toHaveBeenLastCalledWith("12")
   })
 
-  test('input type should be "password" when mask is true', async () => {
-    await render(<PinInput.Root mask />)
-
-    const inputs = page.getByPlaceholder("◯").elements() as HTMLInputElement[]
-
-    inputs.forEach((input) => {
-      expect(input).toHaveAttribute("type", "password")
-    })
-  })
-
-  test("correctly applies custom placeholder to each input", async () => {
-    const customPlaceholder = "t"
-    const { container } = await render(
-      <PinInput.Root placeholder={customPlaceholder} />,
-    )
-
-    const inputs = container.querySelectorAll(
-      `input[placeholder="${customPlaceholder}"]`,
-    )
-
-    expect(inputs).toHaveLength(4)
-  })
-
-  test('sets autoComplete to "one-time-code" when otp is true', async () => {
-    await render(<PinInput.Root otp />)
-
-    const inputs = getInputs()
-    inputs.forEach((input) => {
-      expect(input).toHaveAttribute("autoComplete", "one-time-code")
-    })
-  })
-
-  test('does not set autoComplete to "one-time-code" when otp is false', async () => {
-    await render(<PinInput.Root otp={false} />)
-
-    const inputs = getInputs()
-    inputs.forEach((input) => {
-      expect(input).not.toHaveAttribute("autoComplete", "one-time-code")
-    })
-  })
-
-  test("correctly sets defaultValue into each input", async () => {
-    const defaultValue = "1234"
-
-    await render(<PinInput.Root defaultValue={defaultValue} />)
-
-    const inputs = getInputs()
-
-    inputs.forEach((input, index) => {
-      expect(input).toHaveValue(defaultValue[index])
-    })
-
-    expect(inputs).toHaveLength(defaultValue.length)
-  })
-
   test("correct behavior on input focus", async () => {
     const { user } = await render(<PinInput.Root />)
 
@@ -205,7 +72,7 @@ describe("<PinInput />", () => {
     expect(secondInput?.placeholder).toBe("")
   })
 
-  test("focus moves to previous input on backspace if current input is empty and manageFocus is true", async () => {
+  test("focus moves to previous input on backspace if current input is empty and `manageFocus` is true", async () => {
     const { user } = await render(
       <PinInput.Root defaultValue="12" manageFocus />,
     )
@@ -214,7 +81,6 @@ describe("<PinInput />", () => {
     const currentInput = inputs[2]
     const isFirefox = /firefox/i.test(navigator.userAgent)
     const expectedFocusedInput = isFirefox ? currentInput : inputs[1]
-    const expectedPreviousInputValue = ""
 
     await user.click(currentInput!)
     await expect.poll(() => document.activeElement).toStrictEqual(currentInput)
@@ -223,10 +89,10 @@ describe("<PinInput />", () => {
     await expect
       .poll(() => document.activeElement)
       .toStrictEqual(expectedFocusedInput)
-    await expect.poll(() => inputs[1]?.value).toBe(expectedPreviousInputValue)
+    await expect.poll(() => inputs[1]?.value).toBe("")
   })
 
-  test("does not move focus on backspace if manageFocus is false", async () => {
+  test("does not move focus on backspace if `manageFocus` is false", async () => {
     const { user } = await render(
       <PinInput.Root defaultValue="12" manageFocus={false} />,
     )
@@ -241,7 +107,7 @@ describe("<PinInput />", () => {
     await expect.poll(() => document.activeElement).toStrictEqual(currentInput)
   })
 
-  test("focus move input on arrowRight or arrowLeft if manageFocus is true", async () => {
+  test("focus moves on `ArrowRight` or `ArrowLeft` if `manageFocus` is true", async () => {
     const { user } = await render(
       <PinInput.Root defaultValue="1234" manageFocus />,
     )
@@ -260,24 +126,12 @@ describe("<PinInput />", () => {
     await expect.poll(() => document.activeElement).toStrictEqual(firstInput)
   })
 
-  test("automatically focuses the first input on mount if autoFocus is true", async () => {
+  test("automatically focuses the first input on mount if `autoFocus` is true", async () => {
     await render(<PinInput.Root autoFocus />)
-
-    await new Promise((resolve) => requestAnimationFrame(resolve))
 
     const [firstInput] = getInputs()
 
     await expect.poll(() => document.activeElement).toStrictEqual(firstInput)
-  })
-
-  test("does not focus the first input on mount if autoFocus is false", async () => {
-    await render(<PinInput.Root autoFocus={false} />)
-
-    await new Promise((resolve) => requestAnimationFrame(resolve))
-
-    const [firstInput] = getInputs()
-
-    await expect.poll(() => document.activeElement === firstInput).toBe(false)
   })
 
   test("correct input behavior when pasting a value of 2 characters", async () => {
